MYSQL configuration issue solved

+++ ApplicationManagement/DbModel/EducationalStatus.cs
@@ -12,8 +12,8 @@ namespace ApplicationManagement.DbModel
         public string ExamName { get; set; }
         public string BoardOrUniversity { get; set; }
         public string GroupOrSubject { get; set; }
-        public UInt16 YearOfPassing { get; set; }
-        public UInt16 YearOfExam { get; set; }
+        public Int64 YearOfPassing { get; set; }
+        public Int64 YearOfExam { get; set; }
         public string DivisionOrClassOrGPAOrCGPA { get; set; }
         public string CertificatePdfFileUrl { get; set; }
         public string TranscriptPdfFileUrl { get; set; }
